Add color options
Everything works in its current state, but the formatting needs a lot of work. Also, some of the more advanced functionality like being able to select between if you want to affect the foreground or the background has yet to be implemented, but the process should not be too challenging.